requirements.txt
Flask
flask-restx
pymongo
python-dotenv
